requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.1
pillow==11.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-decouple==3.8
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32==308
requests==2.32.3
scipy==1.14.1
soupsieve==2.6
SpeechRecognition==3.13.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
